Factor out QF internals into separate file and do not expose them via the interface
