+++ src/gqf_file.c
@@ -32,7 +32,9 @@
 #include <sys/stat.h>
 #include <fcntl.h>
 
+#include "hashutil.h"
 #include "gqf.h"
+#include "gqf_int.h"
 #include "gqf_file.h"
 
 #define NUM_SLOTS_TO_LOCK (1ULL<<16)
